sdk/python/pyproject.toml
[tool.poetry]
name = "trainloop-llm-logging"
version = "0.7.5"
description = "TrainLoop LLM Logging SDK for data collection"
authors = ["Mason Pierce <masonpierce@trainloop.ai>"]
readme = "README.md"
packages = [{ include = "trainloop_llm_logging" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pyyaml = "^6.0"
fsspec = ">=2023.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
pytest-timeout = "^2.2.0"
pytest-xdist = "^3.5.0"
freezegun = "^1.4.0"
responses = "^0.24.1"
black = "^23.3.0"
mypy = "^1.5.0"
httpx = "^0.28.1"
requests = "^2.32.3"
python-dotenv = "^1.0.0"
# LLM libraries for integration testing
litellm = "^1.50.0"
openai = "^1.50.0"
anthropic = "^0.39.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
instructor = "^1.6.0"
pydantic = "^2.9.0"
browser-use = "^0.1.0"
playwright = "^1.40.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
